fix(b04-transactions-balance): grader unknown-void, bigint money locals, signup-bonus ledger, FK/index hardening
Settlement/balance correctness batch (9 findings). DB NOT touched — sources
only (apply via the apply_*.ts scripts / migration review, manually after review).

04.1 (HIGH) fixed_odds_grade: a completed bout with a winner but method=null
  (bucket 'unknown', method scraped later than the result) now grades winner-only
  and VOIDs method/totals/distance, mirroring settleSelection. Previously method
  codes fell through to 'lost' and totals/distance graded off a false
  went_distance / raw round, paying or busting real-coin props.
  [apply_parlay_settlement.ts, 0088]

04.2/04.3 (HIGH/MED) bigint money-path locals: LMSR settle_market_* + refund_market
  + unlock_achievement + check_and_promote_tier widened from int4 to bigint
  (balances/amount/balance_after are bigint since Wave 58), incl. ROUND(shares)
  and floor(stake*odds) casts -> ::bigint so the cast no longer caps the value.
  [apply_notification_params.ts, 0065, 0088]

04.4 (MED) signup bonus ledger: handle_new_user now writes the opening
  10,000-coin 'signup_bonus' transaction (RETURNING the new profile id). The
  balance is NOT bumped again — the row only mirrors the column-default grant so
  the ledger reconciles from account creation. [0055]

04.5 (LOW) parlay payout: a fully-won parlay with no voided leg now pays the
  slip's stored potential_payout instead of re-deriving via exp(sum(ln(float4)))
  (which drifted a coin at the rounding boundary); re-prices only when a leg
  voided & dropped out. [apply_parlay_settlement.ts, apply_notification_params.ts]

04.6 (LOW) canonical consistency: every copy of settle_parlay_legs_for_bout /
  settle_fixed_odds_bets_for_bout / balance fns is now bigint + carries the same
  fixes, so no re-run of either apply script (or migration replay) can reintroduce
  the int4 / always-recompute version. [0085 -> bigint; 0088 supersedes 0086_wave59]

04.7 (LOW) RLS / auth-delete trigger: no action (deferred to Wave 3 per project
  decision) — noted only.

04.8 (LOW) bet.outcomeId -> onDelete: cascade (bets die with their outcome,
  consistent with the market->bet / market->market_outcome cascades; cascade not
  restrict, which would fire immediately and break the market-delete diamond).
  [markets.ts]

04.9 (LOW) composite (bout_id, status) indexes on fixed_odds_bet & parlay_leg to
  serve the settlement predicate (WHERE bout_id=X AND status='open') directly.
  [markets.ts]

New migration: 0088_wave60_settlement_balance_bigint.sql — records the LMSR
bigint widening + the grading helpers (with the 04.1 void fix) + the parlay
stored-payout fix in migration history; bodies kept byte-identical to the live
apply_*.ts scripts.

Files outside the assigned B04 set (minimal, additive — flagged per batch rules):
  - src/lib/sportsbook.test.ts: added a describe block locking the 04.1 reference
    behaviour (completed+winner+null-method -> winner-only/void) that the SQL
    grader mirrors. No existing tests changed.

Verified: tsc --noEmit clean; 92/92 unit tests pass (incl. the 2 new cases);
eslint clean on changed files; adversarial review of grader-equivalence,
bigint-completeness, parlay-payout-consistency and signup/schema all clear.

+++ drizzle/migrations/0065_wave47_username_and_tier.sql
@@ -10,6 +10,13 @@
 --   gold      200,000 → 1,000
 --   diamond   500,000 → 1,500
 --   champion  1,000,000 → 2,000
+--
+-- Balance/payout locals are declared bigint here: balance_coins /
+-- total_coins_earned / transaction.balance_after were widened to bigint
+-- (Wave 58), so `RETURNING balance_coins INTO v_new_balance` (and ROUND(shares)
+-- → v_payout) overflows / raises `integer out of range` on an int4 local once a
+-- balance crosses ~2.1e9. Mirrors the Wave-59 widening of the fixed-odds/parlay
+-- path (0086_wave59) for the LMSR + tier/achievement functions.
 
 ALTER TABLE user_profile
   ADD COLUMN IF NOT EXISTS username_last_changed_at timestamptz;
@@ -25,7 +32,7 @@ SECURITY DEFINER
 SET search_path = public
 AS $$
 DECLARE
-  v_total_earned int;
+  v_total_earned bigint;
   v_current_tier text;
   v_new_tier text;
   v_username text;
@@ -80,7 +87,7 @@ AS $$
 DECLARE
   v_achievement_id uuid;
   v_reward int;
-  v_new_balance int;
+  v_new_balance bigint;
   v_name text;
   v_description text;
   v_username text;
@@ -157,8 +164,8 @@ DECLARE
   v_losing_id uuid;
   v_winning_label text;
   v_bet record;
-  v_payout int;
-  v_new_balance int;
+  v_payout bigint;
+  v_new_balance bigint;
 BEGIN
   IF p_winning_idx NOT IN (0, 1) THEN
     RAISE EXCEPTION 'winning_idx must be 0 or 1, got %', p_winning_idx;
@@ -199,7 +206,7 @@ BEGIN
       AND outcome_id = v_winning_id
       AND resolved_at IS NULL
   LOOP
-    v_payout := ROUND(v_bet.shares_bought)::int;
+    v_payout := ROUND(v_bet.shares_bought)::bigint;
 
     UPDATE bet
       SET payout = v_payout, resolved_at = NOW()
@@ -265,8 +272,8 @@ DECLARE
   v_winning_id uuid;
   v_winning_label text;
   v_bet record;
-  v_payout int;
-  v_new_balance int;
+  v_payout bigint;
+  v_new_balance bigint;
 BEGIN
   IF p_winner_offset NOT IN (0, 3) THEN
     RAISE EXCEPTION 'winner_offset must be 0 or 3, got %', p_winner_offset;
@@ -309,7 +316,7 @@ BEGIN
       AND outcome_id = v_winning_id
       AND resolved_at IS NULL
   LOOP
-    v_payout := ROUND(v_bet.shares_bought)::int;
+    v_payout := ROUND(v_bet.shares_bought)::bigint;
 
     UPDATE bet
       SET payout = v_payout, resolved_at = NOW()
